+++ packages/generator/src/v1.1/backcompat/internal-foundry-ontology-runtime-dist/baseTypes/generateObjectSetDir.ts
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2023 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import path from "node:path";
+import type { MinimalFs } from "../../../../MinimalFs";
+import { formatTs } from "../../../../util/test/formatTs";
+import { reexportTypes } from "../../util/reexportTypes";
+
+export async function generateObjectSetDir(
+  objectSetDir: string,
+  fs: MinimalFs,
+) {
+  await fs.mkdir(objectSetDir, { recursive: true });
+
+  await fs.writeFile(
+    path.join(objectSetDir, "index.ts"),
+    await formatTs(`
+    export * from "./ObjectSetDefinition";
+    export * from "./OntologyObjectSet";
+    `),
+  );
+
+  await fs.writeFile(
+    path.join(objectSetDir, "ObjectSetDefinition.ts"),
+    await formatTs(
+      reexportTypes([
+        "BaseObjectSetDefinition",
+        "ReferenceObjectSetDefinition",
+        "StaticObjectSetDefinition",
+        "IntersectObjectSetDefinition",
+        "SubtractObjectSetDefinition",
+        "SearchAroundObjectSetDefinition",
+        "FilterObjectSetDefinition",
+        "ObjectSetDefinition",
+      ]),
+    ),
+  );
+
+  await fs.writeFile(
+    path.join(objectSetDir, "OntologyObjectSet.ts"),
+    await formatTs(
+      reexportTypes([
+        "OntologyObjectSet",
+      ]),
+    ),
+  );
+}