+++ packages/legacy-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/legacy-client
 
+## 0.0.16
+
+### Patch Changes
+
+- b59cb10: Adding backcompatibility for internal types
+
 ## 0.0.15
 
 ### Patch Changes